requirements.txt
openai
langchain-openai
langchain-experimental
crewai
crewai[tools]
tiktoken
PyPDF2
pyautogen
pandas
flask